+++ Panther/include/Context.h
@@ -347,6 +347,7 @@ namespace pcit::panther{
 				TypeInfoID typeID;
 
 				bool allowedInConstexpr;
+				bool allowedInComptime;
 				bool allowedInRuntime;
 
 				bool allowedInCompile;
@@ -362,13 +363,15 @@ namespace pcit::panther{
 					evo::Variant<TypeInfo::ID, uint32_t> type; // uint32_t is the index of the templateParam
 				};
 
-				using ReturnParam = evo::Variant<TypeInfo::VoidableID, uint32_t>; // uint32_t is the index of the templateParam
+				using ReturnParam = evo::Variant<TypeInfo::VoidableID, uint32_t>; // uint32_t is the index 
+																				  //   of the templateParam
 
 				evo::SmallVector<std::optional<TypeInfo::ID>> templateParams; // nullopt means it's a `Type` param
 				evo::SmallVector<Param> params;
 				evo::SmallVector<ReturnParam> returns;
 
 				bool allowedInConstexpr;
+				bool allowedInComptime;
 				bool allowedInRuntime;
 
 				bool allowedInCompile;
@@ -383,7 +386,7 @@ namespace pcit::panther{
 					instantiated_params.reserve(this->params.size());
 					for(const Param& param : this->params){
 						const TypeInfo::ID param_type = param.type.visit([&](const auto& param_type) -> TypeInfo::ID {
-							if constexpr(std::is_same_v<std::decay_t<decltype(param_type)>, TypeInfo::ID>){
+							if constexpr(std::is_same<std::decay_t<decltype(param_type)>, TypeInfo::ID>()){
 								return param_type;
 							}else{
 								return template_args[param_type]->asTypeID();
@@ -397,7 +400,7 @@ namespace pcit::panther{
 					instantiated_returns.reserve(this->returns.size());
 					for(const ReturnParam& return_param : this->returns){
 						const TypeInfo::VoidableID return_type = return_param.visit([&](const auto& return_data){
-							if constexpr(std::is_same_v<std::decay_t<decltype(return_data)>, TypeInfo::VoidableID>){
+							if constexpr(std::is_same<std::decay_t<decltype(return_data)>, TypeInfo::VoidableID>()){
 								return return_data;
 							}else{
 								return *template_args[return_data];