+++ Panther/src/sema_to_pir/SemaToPIR.h
@@ -62,6 +62,11 @@ namespace pcit::panther{
 			auto get_expr_impl(const sema::Expr expr, evo::ArrayProxy<pir::Expr> store_locations)
 				-> std::optional<pir::Expr>;
 
+			template<GetExprMode MODE>
+			auto template_intrinsic_func_call(
+				const sema::FuncCall& func_call, evo::ArrayProxy<pir::Expr> store_locations
+			) -> std::optional<pir::Expr>;
+
 			auto intrinsic_func_call(const sema::FuncCall& func_call) -> void;
 
 			EVO_NODISCARD auto get_global_var_value(const sema::Expr expr) -> pir::GlobalVar::Value;